Function to create JSON serialized output
Create a `serializeJSON()` function to output the current mertrics to a file
in JSON format. This uses a basic JSON formatting where each metric is listed
as a field under the top level metrics document.

The families of metrics are respected, allowing users to still create multiple
metrics in a single family. The prometheus fields of type and help are
disregarded in JSON serializing.