Document library and create example program
Create a full documentation set for the library, indicating the usage for
the main metrics types. Documentation has also been created for the library
description, requirements, build instruction, and artifact output.

An example program was created to demonstrate how to use the library as
a full example. A build script was also created for this example program
so it can be compiled and ran as a demonstration. This also acts as
instruction on how to link in the compiled metrics library